requirements.txt
anthropic
fastapi
uvicorn[standard]
google-auth-oauthlib
google-api-python-client
python-dotenv
pydantic
apscheduler
